Update README.md
ABOUT THE PROJECT:
KNOW THE BLOCKS MAVEN is an Educational Consult passionate about educating and training YOUNG Nigerians about the Blockchain ecosystem. 
Know the blocks maven offers expertise in Blockchain Ecosystem Training. Our tried and true system is based off on several years’ cumulative experience shared between our trainers. We pride ourselves on setting up our Students for success in the Blockchain industry and are sure that you will leave our sessions more prepared, informed and knowledgeable with deeper insights into various careers available to explore in today's train.
OUR MOTIVATION
Blockchain is an evolving technology that has come to stay and will grow its tentacles to every industry in the world. Since this is an evolving technology expanding with the speed of light, a great need has arisen globally for experts in different careers to offer their services in the Blockchain Ecosystem. And guess what? We have very little of them as of yet. The career paths in the Blockchain Ecosystem is grossly under-explored and this is where we come in. We enlighten, teach, train and practically hold these young Nigerians by the hand to becoming fully-fledged experts in whatever career they choose in the Blockchain Ecosystem.
VISION AND MISSION
Our vision for this project is born out of a passion for solving the enormous challenge of unemployment in Nigeria by looking at other niches in the career world that haven’t been “oversaturated” and yet will remain relevant in the next multiple decades. 
Ninety per cent out of a Hundred when asked what do you know about the blockchain will mention cryptocurrency, that is the Bitcoin and the Ethereum. That is one of the gaps we want to bridge. INFORMATION. Blockchain is a database system hence, its numerous application to other major sectors of the world like the healthcare, real estate, governance etc. is becoming widely accepted in the world. Its importance in the next decade cannot be overemphasized.
HOW CAN KNOW THE BLOCKS MAVEN HELP? 
•	Know The Blocks Maven imparts world-class training both onsite & online. 
•	We believe cohort-based collaborative learning helps people learn faster. Self-learners still need shared environments to practice their skills. 
•	We are experts at giving you all these in the blockchain space which will help you gain a competitive edge in the blockchain career niche. 
•	We aim to impact this knowledge to early teens and youth by integrating and familiarizing them with the Blockchain Ecosystem.
WHAT ARE OUR TRAINING SESSIONS LIKE?
Our training sessions are meticulously planned by our trainer and are designed in a manner geared toward the maximization of efficiency. Your time is valuable, and so is ours, so our goal is to best prepare you while taking up the least amount of your time possible.
Our training will be onsite at the above address and will be conducted by our trainers from different parts of the world with a total of sixteen (16) sessions. Training sessions will be comprised of both activities and lectures that will touch upon various subjects after an introductory session that enlightens each trainee on their preferred career paths in the Blockchain which can either be onsite or online.